Improve message grammar

+++ core/src/main/java/de/sayayi/lib/message/data/map/MapKeyNumber.java
@@ -32,7 +32,7 @@ import java.util.Locale;
 public final class MapKeyNumber implements MapKey
 {
   private CompareType compareType;
-  @Getter private final int number;
+  @Getter private final long number;
 
 
   @NotNull
